refactor: Split transaction module
